CCLabelTTF works ok
Conflicts:

	cocos2d/CCLabelTTF.h
	cocos2d/CCTexture2D.h
	cocos2d/CCTexture2D.m
	mac/cocos2d-mac.xcodeproj/project.pbxproj


Former-commit-id: 952af89ab72bd2924c2d572f9af4ffb6c13d2562

+++ cocos2d/Support/ccCArray.h
@@ -49,7 +49,7 @@
 
 // Easy integration	
 #define CCARRAYDATA_FOREACH(__array__, __object__)															\
-__object__=__array__->arr[0]; for(int i=0, num=__array__->num; i<num; i++, __object__=__array__->arr[i])	\
+__object__=__array__->arr[0]; for(NSUInteger i=0, num=__array__->num; i<num; i++, __object__=__array__->arr[i])	\
 
 
 typedef struct ccArray {
